+++ server/routes/collection.js
@@ -11,12 +11,12 @@ router.post('/free', async (req, res, next) => {
     if (result) {
         return res.send({
             code: 200,
-            'status': 'SUCCESS'
+            status: 'SUCCESS'
         })
     } else {
         return res.send({
             code: 500,
-            'status': 'SERVER ERROR'
+            status: 'SERVER ERROR'
         })
     }
 })
