+++ server/routes/collections.js
@@ -20,12 +20,4 @@ router.get('/collections_free_fromUsers', async (req, res) => {
     res.send(rows);
 });
 
-router.post('/collections_free', async (req, res) => {
-    console.log(req.body);
-    const {collection_name, collection_private, collection_keywords, collection_type} = req.body;
-    //회원가입 때 날아온 user_pk를 넣어줘야한다
-    const result = await db.query('insert into collections(collection_name, user_pk, collection_private, collection_keywords, collection_type) values (?, 3, ?, ?, ?);', [collection_name, collection_private, collection_keywords, collection_type]);
-    console.log(result)
-})
-
 module.exports = router;
