+++ server/routes/index.js
@@ -4,18 +4,16 @@ const router = express.Router();
 // 라우터 파일 가져오기
 const collectionRouter = require('./collection');
 const authRouter = require('./auth');
-const keywordRouter = require('./keywords');
 const userRouter = require('./users');
-const tourApiRouter = require('./tourApi');
 const searchRouter = require('./search');
+const placeRouter = require('./place');
 
 
 // 라우터 파일 등록하기
-router.use('/collection', collectionRouter);
 router.use('/auth', authRouter);
+router.use('/collection', collectionRouter);
 router.use('/user', userRouter);
-router.use('/keyword', keywordRouter);
-router.use('/tourapi', tourApiRouter);
 router.use('/search', searchRouter);
+router.use('/place', placeRouter);
 
 module.exports = router;
