+++ server/services/authService.js
@@ -1,13 +1,13 @@
 const db = require('../database/database')
 const mysql = require('mysql2')
 
-exports.addUser = async (userData) => {
+exports.createUser = async (userData) => {
     const query = `INSERT INTO users SET ?`;
     const result = await db.query(query, userData);
     return result;
 }
 
-exports.findUser = async (email) => {
+exports.readUserCnt = async (email) => {
     const query = `SELECT COUNT(user_email) 
                    AS count 
                    FROM users 
@@ -16,7 +16,7 @@ exports.findUser = async (email) => {
     return result;
 }
 
-exports.getUser = async (email) => {
+exports.readUser = async (email) => {
     const query = `SELECT *
                    FROM users
                    WHERE user_email=${mysql.escape(email)}`
