+++ server/services/placeService.js
@@ -2,7 +2,7 @@ const db = require('../database/database');
 const mysql = require('mysql2');
 
 // 장소 배열 조회
-exports.selectPlaceList = async (keyword) => {
+exports.readPlaceList = async (keyword) => {
     const query = `SELECT place_pk, place_name, place_addr, place_img, place_type
                    FROM places
                    WHERE place_name LIKE ${mysql.escape(`%${keyword}%`)}`;
@@ -13,5 +13,14 @@ exports.selectPlaceList = async (keyword) => {
         return {...place, star: 3}
     })
 
+    return result;
+}
+
+exports.readPlace = async (place_pk) => {
+    const query = `SELECT place_pk, place_name, place_addr, place_type 
+                   FROM places 
+                   WHERE place_pk=${place_pk}`
+
+    const result = await db.query(query);
     return result;
 }
